[PACKAGE] : 클린 아키텍처 패키지 네이밍 수정
