--- src/main/java/kr/hhplus/be/server/controller/user/dto/response/UserPointResponse.java
@@ -1,6 +1,0 @@
-package kr.hhplus.be.server.controller.user.dto.response;
-
-public record UserPointResponse(
-        long amount
-) {
-}
